Major restructure and clean up
