+++ pysweep.yaml
@@ -1,0 +1,19 @@
+---
+  blocksize: 32
+  share: 0.5
+    0
+  time:
+    t0: 0
+    tf: 1
+    dt: 0.1
+
+  global:
+    gamma: 1.4
+
+  swept:
+    intermediate_steps: 2
+    operating_points: 2
+
+  source:
+    cpu:
+    gpu:
