+++ pysweep/__init__.py
@@ -4,3 +4,4 @@ from . import sweep
 from . import equations
 from . import tests
 from . import commandline
+from . import gputools
