+++ pysweep/core/equations/LxF.py
