+++ pysweep/core/equations/__init__.py
