+++ pysweep/core/equations/eqt.cu
