+++ pysweep/core/equations/euler.cu
