+++ pysweep/core/equations/euler.py
