+++ pysweep/core/equations/hde.py
