+++ pysweep/core/equations/test_euler.py
