+++ pysweep/core/interface.py
@@ -1,0 +1,2 @@
+import ctypes, os
+path = os.path.dirname(os.path.abspath(__file__))
