--- pysweep/decomposition/__init__.py
@@ -1,4 +1,0 @@
-from . import ddecomp
-from . import dece
-from . import ddcore
-from . import ccore
