--- pysweep/decomposition/ccore/__init__.py
@@ -1,2 +1,0 @@
-from . import printer
-from . import source
