--- pysweep/decomposition/ddcore/__init__.py
@@ -1,5 +1,0 @@
-from . import block
-from . import dcore
-from . import decomp
-from . import functions
-from . import sgs
