+++ pysweep/dse.py
@@ -1,23 +1,14 @@
 #Programmer: Anthony Walker
 #PySweep is a package used to implement the swept rule for solving PDEs
-import sys, os, h5py, math, GPUtil, time
-from itertools import cycle
+import sys, os, h5py, numpy, time, pickle, itertools.cycle as cycle, multiprocessing as mp, mpi4py.MPI as MPI
+from core import core, block, functions, sgs, interface, io, process, GPUtil
 #CUDA Imports
 try:
     import pycuda.driver as cuda
 except Exception as e:
     print(str(e)+": Importing pycuda failed, execution will continue but is most likely to fail unless the affinity is 0.")
-#Dsweep imports
-from ddcore import dcore,decomp,functions,sgs
-from ccore import source, printer
-#MPI imports
-from mpi4py import MPI
-#Multi-processing imports
-import multiprocessing as mp
-import numpy as np
 
-def decomp_engine():
-    # arr0,gargs,swargs,filename ="results",exid=[],dType=np.dtype('float32')
+def dsweep_engine():
     """Use this function to perform swept rule
     args:
     arr0 -  3D numpy array of initial conditions (v (variables), x,y)
@@ -33,13 +24,12 @@ def decomp_engine():
         AF -  the GPU affinity (GPU work/CPU work)/TotalWork
         GS - GPU source code file
         CS - CPU source code file
-    dType: data type for the run (e.g. np.dtype("float32")).
     filename: Name of the output file excluding hdf5
     exid: GPU ids to exclude from the calculation.
     """
     #Starting timer
     start = time.time()
-    #Setting global variables
+    #initialize global cpu arr
     sgs.init_globals()
     #Local Constants
     ZERO = 0
@@ -51,11 +41,18 @@ def decomp_engine():
     #------------------INPUT DATA SETUP-------------------------$
     arr0,gargs,swargs,filename,exid,dType = decomp.read_input_file(comm)
     TSO,OPS,BS,AF = [int(x) for x in swargs[:-1]]+[swargs[-1]]
-    sgs.TSO,sgs.OPS,sgs.BS,sgs.AF = [int(x) for x in swargs[:-1]]+[swargs[-1]]
     t0,tf,dt = gargs[:3]
     assert BS%(2*OPS)==0, "Invalid blocksize, blocksize must satisfy BS = 2*ops*k and the architectural limit where k is any integer factor."
     BS = (BS,BS,1)
+    #---------------------SWEPT VARIABLE SETUP----------------------$
+    #Splits for shared array
+    SPLITX = int(BS[ZERO]/TWO)   #Split computation shift - add OPS
+    SPLITY = int(BS[ONE]/TWO)   #Split computation shift
+    MPSS = int(BS[0]/(2*OPS)-1)
+    MOSS = 2*MPSS
     time_steps = int((tf-t0)/dt)  #Number of time steps
+    MGST = int(TSO*(time_steps-MPSS)/(MPSS)+1)  #Global swept step  #THIS ASSUMES THAT time_steps > MOSS
+    time_steps = int(MPSS*(MGST+1)/TSO+1) #Number of time steps - Add 1 for initial conditions
     #-------------MPI SETUP----------------------------#
     processor = MPI.Get_processor_name()
     rank = comm.Get_rank()  #current rank
@@ -82,13 +79,13 @@ def decomp_engine():
     if NMB:
         #Giving each node an id
         if cluster_master == rank:
-            node_id = np.arange(1,cluster_comm.Get_size()+1,1,dtype=np.intc)
+            node_id = numpy.arange(1,cluster_comm.Get_size()+1,1,dtype=numpy.intc)
         else:
             node_id = None
         node_id = cluster_comm.scatter(node_id)
         #Getting GPU information
         gpu_rank,total_num_gpus, num_gpus, node_info, comranks, GNR,CNR = dcore.get_gpu_info(rank,cluster_master,node_id,cluster_comm,AF,BS,exid,processors,node_comm.Get_size(),arr0.shape)
-        ranks_to_remove = dcore.find_remove_ranks(node_ranks,AF,num_gpus)
+        ranks_to_remove = dcore.find_remove_ranks(node_ranks,AF,num_gpus,CNR)
         [gpu_rank.append(None) for i in range(len(node_ranks)-len(gpu_rank))]
         #Testing ranks and number of gpus to ensure simulation is viable
         assert total_num_gpus < comm.Get_size() if AF < 1 else True,"The affinity specifies use of heterogeneous system but number of GPUs exceeds number of specified ranks."
@@ -103,83 +100,75 @@ def decomp_engine():
     node_info = node_comm.bcast(node_info)
     #----------------------__Removing Unwanted MPI Processes------------------------#
     node_comm,comm = dcore.mpi_destruction(rank,node_ranks,comm,ranks_to_remove,all_ranks)
-    gpu_rank,blocks = decomp.nsplit(rank,node_master,node_comm,num_gpus,node_info,BS,arr0.shape,gpu_rank,OPS)
+    gpu_rank,blocks = decomp.nsplit(rank,node_master,node_comm,num_gpus,node_info,BS,arr0.shape,gpu_rank)
     # Checking to ensure that there are enough
     assert total_num_gpus >= node_comm.Get_size() if AF == 1 else True,"Not enough GPUs for ranks"
     #---------------------------Creating and Filling Shared Array-------------#
-    shared_shape = (TSO+1,arr0.shape[0],int(sum(node_info[2:])*BS[0]+2*OPS),arr0.shape[2]+2*OPS)
-    sarr = decomp.create_CPU_sarray(node_comm,shared_shape,dType,np.zeros(shared_shape).nbytes)
-    #Making blocks match array other dimensions
-    bsls = [slice(0,i,1) for i in shared_shape]
-    blocks = (bsls[0],bsls[1],blocks,slice(bsls[3].start+OPS,bsls[3].stop-OPS,1))
-    GRB = True if gpu_rank is not None else False
+    shared_shape = (MOSS+TSO+ONE,arr0.shape[0],int(sum(node_info[2:])*BS[0]),arr0.shape[2])
+    sarr = decomp.create_CPU_sarray(node_comm,shared_shape,dType,numpy.zeros(shared_shape).nbytes)
     #Filling shared array
     if NMB:
-        gsc = (slice(0,arr0.shape[0],1),slice(int(node_info[0]*BS[0]),int(node_info[1]*BS[0]),1),slice(0,arr0.shape[2],1))
-        i1,i2,i3 = gsc
-        #TSO STEP and BOUNDARIES
-        sarr[TSO-ONE,:,OPS:-OPS,OPS:-OPS] =  arr0[gsc]
-        sarr[TSO-ONE,:,OPS:-OPS,:OPS] = arr0[gsc[0],gsc[1],-OPS-1:-1]
-        sarr[TSO-ONE,:,OPS:-OPS,-OPS:] = arr0[gsc[0],gsc[1],1:OPS+1]
-        #INITIAL STEP AND BOUNDARIES
-        sarr[0,:,OPS:-OPS,OPS:-OPS] =  arr0[gsc]
-        sarr[0,:,OPS:-OPS,:OPS] = arr0[gsc[0],gsc[1],-OPS-1:-1]
-        sarr[0,:,OPS:-OPS,-OPS:] = arr0[gsc[0],gsc[1],1:OPS+1]
+        gsc = (slice(0,arr0.shape[1],1),slice(int(node_info[0]*BS[0]),int(node_info[1]*BS[0]),1),slice(0,arr0.shape[2],1))
+        sarr[TSO-ONE,:,:,:] =  arr0[gsc]
+        sarr[0,:,:,:] =  arr0[gsc]
     else:
         gsc = None
+
+    #Making blocks match array other dimensions
+    bsls = [slice(0,i,1) for i in shared_shape]
+    blocks = (bsls[0],bsls[1],blocks,bsls[3])
+    GRB = True if gpu_rank is not None else False
     # ------------------- Operations specifically for GPus and CPUs------------------------#
     if GRB:
         # Creating cuda device and context
         cuda.init()
         cuda_device = cuda.Device(gpu_rank)
         cuda_context = cuda_device.make_context()
-        DecompObj,garr = dcore.gpu_core(blocks,BS,OPS,gargs,GRB,TSO)
-        mpi_pool= None
+        SM,garr,Up,Down,Oct,Xb,Yb  = dcore.gpu_core(blocks,BS,OPS,gargs,GRB,MPSS,MOSS,TSO)
+        mpiPool,total_cpu_block = None,None
     else:
-        garr = None
-        DecompObj,blocks= dcore.cpu_core(sarr,blocks,shared_shape,OPS,BS,GRB,TSO,gargs)
-        pool_size = min(len(blocks), os.cpu_count()-node_comm.Get_size()+1)
-        mpi_pool = mp.Pool(pool_size)
-    functions.send_edges(sarr,NMB,GRB,node_comm,cluster_comm,comranks,OPS,garr,DecompObj)
+        SM,garr = None,None
+        blocks,total_cpu_block,Up,Down,Oct,Xb,Yb = dcore.cpu_core(sarr,blocks,shared_shape,OPS,BS,gargs,MPSS,TSO)
+        poolSize = min(len(blocks[0]), os.cpu_count()-node_comm.Get_size()+1)
+        mpiPool = mp.Pool(poolSize)
+
     # ------------------------------HDF5 File------------------------------------------#
     hdf5_file, hdf5_data,hdf_time = dcore.make_hdf5(filename,cluster_master,comm,rank,BS,arr0,time_steps,AF,dType,gargs)
     comm.Barrier() #Ensure all processes are prepared to solve
-    # -------------------------------Standard Decomposition---------------------------------------------#
+    stop1 = time.time()
+    # print(stop1-start)
+    # -------------------------------SWEPT RULE---------------------------------------------#
+    # -------------------------------FIRST PRISM AND COMMUNICATION-------------------------------------------#
+    functions.FirstPrism(SM,GRB,Up,Yb,mpiPool,blocks,sarr,garr,total_cpu_block)
     node_comm.Barrier()
-    cwt = 1
-    # if rank == node_master:
-    #     for currt in range(0,1,1):
-    #         print("-------------------",currt,"-------------------")
-    #         for x in sarr[currt,0,:,:]:
-    #                 sys.stdout.write('[')
-    #                 [sys.stdout.write("%d"%cf+", ") for cf in x ]
-    #                 sys.stdout.write(']\n')
-    #         input()
-    # node_comm.Barrier()
-    for i in range(TSO*time_steps):
-        functions.Decomposition(GRB,OPS,sarr,garr,blocks,mpi_pool,DecompObj)
+    functions.first_forward(NMB,GRB,node_comm,cluster_comm,comranks,sarr,SPLITX,total_cpu_block)
+    #Loop variables
+    cwt = 1 #Current write time
+    gts = 0 #Initialization of global time step
+    del Up #Deleting Up object after FirstPrism
+    #-------------------------------SWEPT LOOP--------------------------------------------#
+    step = cycle([functions.send_backward,functions.send_forward])
+    for i in range(MGST):
+        functions.UpPrism(GRB,Xb,Yb,Oct,mpiPool,blocks,sarr,garr,total_cpu_block)
         node_comm.Barrier()
-        #Write data and copy down a step
-        if (i+1)%TSO==0 and NMB:
-            hdf5_data[cwt,i1,i2,i3] = sarr[TSO,:,OPS:-OPS,OPS:-OPS]
-            sarr = np.roll(sarr,TSO,axis=0) #Copy down
-            cwt+=1
-        elif NMB:
-            sarr = np.roll(sarr,TSO,axis=0) #Copy down
-        node_comm.Barrier()
-        #Communicate
-        functions.send_edges(sarr,NMB,GRB,node_comm,cluster_comm,comranks,OPS,garr,DecompObj)
-
+        cwt = next(step)(cwt,sarr,hdf5_data,gsc,NMB,GRB,node_comm,cluster_comm,comranks,SPLITX,gts,TSO,MPSS,total_cpu_block)
+        gts+=MPSS
+    #Do LastPrism Here then Write all of the remaining data
+    Down.gts = Oct.gts
+    functions.LastPrism(GRB,Xb,Down,mpiPool,blocks,sarr,garr,total_cpu_block)
+    node_comm.Barrier()
+    next(step)(cwt,sarr,hdf5_data,gsc,NMB,GRB,node_comm,cluster_comm,comranks,SPLITX,gts,TSO,MPSS,total_cpu_block)
     # Clean Up - Pop Cuda Contexts and Close Pool
     if GRB:
         cuda_context.pop()
     comm.Barrier()
     stop = time.time()
+    # print(stop-stop1)
     hdf_time[0] = stop-start
     hdf5_file.close()
     #Removing input file.
     if rank==cluster_master:
-        # os.system("rm "+"input_file.hdf5") #remove input file
+        os.system("rm "+"input_file.hdf5") #remove input file
         gargs+=swargs[:4]
         if os.path.isfile('log.hdf5'):
             log_file = h5py.File('log.hdf5','a')
@@ -187,18 +176,19 @@ def decomp_engine():
             log_file['time'].resize((log_file['time'].shape[0]+1),axis=0)
             log_file['time'][shape]=stop-start
             log_file['type'].resize((log_file['type'].shape[0]+1),axis=0)
-            log_file['type'][shape]=ord('d')
+            log_file['type'][shape]=ord('s')
             log_file['args'].resize((log_file['args'].shape[0]+1),axis=0)
             log_file['args'][shape]=gargs
             log_file.close()
         else:
             log_file = h5py.File('log.hdf5','w')
             log_file.create_dataset('time',(1,),data=(stop-start),chunks=True,maxshape=(None,))
-            log_file.create_dataset('type',(1,),data=(ord('d')),chunks=True,maxshape=(None,))
-            log_file.create_dataset('args',(1,)+np.shape(gargs),data=gargs,chunks=True,maxshape=(None,)+np.shape(gargs))
+            log_file.create_dataset('type',(1,),data=(ord('s')),chunks=True,maxshape=(None,))
+            log_file.create_dataset('args',(1,)+numpy.shape(gargs),data=gargs,chunks=True,maxshape=(None,)+numpy.shape(gargs))
             log_file.close()
 
 #Statement to execute dsweep
-decomp_engine()
+dsweep_engine()
+
 #Statement to finalize MPI processes
 MPI.Finalize()
