--- pysweep/sweep/__init__.py
@@ -1,4 +1,0 @@
-from . import distsweep
-from . import dse
-from . import dcore
-from . import ccore
