--- pysweep/sweep/dcore/sgs.py
@@ -1,7 +1,0 @@
-#Programmer: Anthony Walker
-#This file is for global variable initialization
-def init_globals():
-    """Use this function to initialize globals."""
-    #Setting global variables
-    global carr
-    carr = None
